+++ sample/src/main/java/com/github/amlcurran/showcaseview/sample/EventsActivity.java
@@ -15,10 +15,6 @@ import com.github.amlcurran.showcaseview.ShowcaseView;
 import com.github.amlcurran.showcaseview.SimpleShowcaseEventListener;
 import com.github.amlcurran.showcaseview.targets.ViewTarget;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 public class EventsActivity extends AppCompatActivity {
 
     @Override
@@ -29,13 +25,14 @@ public class EventsActivity extends AppCompatActivity {
         TextView eventLog = (TextView) findViewById(R.id.events_log);
         Button customButton = (Button) getLayoutInflater().inflate(R.layout.view_custom_button, null);
 
+        MultiEventListener multiEventListener = new MultiEventListener(new LogToTextListener(eventLog), new ShakeButtonListener(customButton));
         new ShowcaseView.Builder(this)
                 .withMaterialShowcase()
                 .setStyle(R.style.CustomShowcaseTheme3)
                 .setTarget(new ViewTarget(R.id.imageView, this))
                 .setContentTitle("Events")
                 .setContentText("Listening to ShowcaseView events is easy!")
-                .setShowcaseEventListener(new MultiEventListener(new LogToTextListener(eventLog), new ShakeButtonListener(customButton)))
+                .setShowcaseEventListener(multiEventListener)
                 .replaceEndButton(customButton)
                 .build();
     }
@@ -75,44 +72,6 @@ public class EventsActivity extends AppCompatActivity {
             append("Touch blocked: x: " + motionEvent.getX() + " y: " + motionEvent.getY());
         }
 
-    }
-
-    private class MultiEventListener implements OnShowcaseEventListener {
-
-        private final List<OnShowcaseEventListener> listeners;
-
-        public MultiEventListener(OnShowcaseEventListener... listeners) {
-            this.listeners = new ArrayList<>();
-            this.listeners.addAll(Arrays.asList(listeners));
-        }
-
-        @Override
-        public void onShowcaseViewHide(ShowcaseView showcaseView) {
-            for (OnShowcaseEventListener listener : listeners) {
-                listener.onShowcaseViewHide(showcaseView);
-            }
-        }
-
-        @Override
-        public void onShowcaseViewDidHide(ShowcaseView showcaseView) {
-            for (OnShowcaseEventListener listener : listeners) {
-                listener.onShowcaseViewDidHide(showcaseView);
-            }
-        }
-
-        @Override
-        public void onShowcaseViewShow(ShowcaseView showcaseView) {
-            for (OnShowcaseEventListener listener : listeners) {
-                listener.onShowcaseViewShow(showcaseView);
-            }
-        }
-
-        @Override
-        public void onShowcaseViewTouchBlocked(MotionEvent motionEvent) {
-            for (OnShowcaseEventListener listener : listeners) {
-                listener.onShowcaseViewTouchBlocked(motionEvent);
-            }
-        }
     }
 
     private class ShakeButtonListener extends SimpleShowcaseEventListener {
